Add startup log and fix the script starting too soon
Terraform doesn't wait for the startup script to complete before doing the remote-exec. Since we're downloading the dependencies in the startup script of the VM at this moment, so we need to wait on that before the script can run. This is now done by simply moving running the script to the startup script code.